Extract DiskIOWorker into separate files.
