+++ src/rendering/caches/DiskCache.cpp
@@ -33,79 +33,6 @@
 
 namespace pag {
 
-// ============================================================================
-// DiskIOWorker Implementation
-// ============================================================================
-
-DiskIOWorker* DiskIOWorker::GetInstance() {
-  // Intentional leak: using *new to avoid static destruction order issues at program exit.
-  // The destructor is implemented correctly but will never be called in this singleton pattern.
-  static auto& instance = *new DiskIOWorker();
-  return &instance;
-}
-
-DiskIOWorker::DiskIOWorker() {
-  workerThread = std::thread(&DiskIOWorker::runLoop, this);
-}
-
-DiskIOWorker::~DiskIOWorker() {
-  {
-    std::lock_guard<std::mutex> lock(locker);
-    stopped = true;
-  }
-  condition.notify_one();
-  if (workerThread.joinable()) {
-    workerThread.join();
-  }
-}
-
-void DiskIOWorker::submit(std::function<void()> task) {
-  if (task == nullptr) {
-    return;
-  }
-  {
-    std::lock_guard<std::mutex> lock(locker);
-    tasks.push(std::move(task));
-  }
-  condition.notify_one();
-}
-
-void DiskIOWorker::waitAll() {
-  std::unique_lock<std::mutex> lock(locker);
-  // Wait until the queue is empty AND no task is currently executing.
-  idleCondition.wait(lock, [this] { return tasks.empty() && !executing; });
-}
-
-void DiskIOWorker::runLoop() {
-  while (true) {
-    std::function<void()> task;
-    {
-      std::unique_lock<std::mutex> lock(locker);
-      condition.wait(lock, [this] { return stopped || !tasks.empty(); });
-      if (stopped && tasks.empty()) {
-        return;
-      }
-      task = std::move(tasks.front());
-      tasks.pop();
-      executing = true;
-    }
-    if (task) {
-      task();
-    }
-    {
-      std::lock_guard<std::mutex> lock(locker);
-      executing = false;
-      if (tasks.empty()) {
-        idleCondition.notify_all();
-      }
-    }
-  }
-}
-
-// ============================================================================
-// DiskCache Implementation
-// ============================================================================
-
 static void RemoveFileTask(std::string path) {
   if (remove(path.c_str()) != 0) {
     LOGE("Failed to remove cached file: %s", path.c_str());
