+++ src/rendering/caches/DiskCache.h
@@ -19,57 +19,15 @@
 #pragma once
 
 #include <atomic>
-#include <condition_variable>
-#include <functional>
 #include <list>
 #include <mutex>
-#include <queue>
-#include <thread>
 #include <unordered_map>
+#include "DiskIOWorker.h"
 #include "SequenceFile.h"
 #include "pag/types.h"
 
 namespace pag {
 class FileInfo;
-
-/**
- * A dedicated background worker thread that executes disk IO tasks serially.
- * Tasks are queued and executed in order on a single thread, avoiding main thread blocking.
- *
- * Lock ordering: DiskIOWorker::locker must be acquired BEFORE DiskCache::locker.
- * IMPORTANT: Never call waitAll() while holding DiskCache::locker, as tasks executed by
- * DiskIOWorker (e.g., CloseFileTask) may acquire DiskCache::locker, causing deadlock.
- */
-class DiskIOWorker {
- public:
-  static DiskIOWorker* GetInstance();
-
-  ~DiskIOWorker();
-
-  /**
-   * Submits a task for asynchronous execution on the background thread.
-   */
-  void submit(std::function<void()> task);
-
-  /**
-   * Waits for all pending tasks to complete. Useful during app shutdown.
-   */
-  void waitAll();
-
- private:
-  DiskIOWorker();
-
-  void runLoop();
-
-  std::mutex locker;
-  std::condition_variable condition;
-  std::condition_variable idleCondition;
-  std::queue<std::function<void()>> tasks;
-  std::thread workerThread;
-  std::atomic<bool> stopped{false};
-  // Tracks if a task is currently being executed. Always accessed under locker, no atomic needed.
-  bool executing = false;
-};
 
 class DiskCache {
  public:
